Remove subfolders (asylum, immigration) for pricing feature tests
Separate immigration from asylum as they have substantial differences.
