+++ lib/submission_config/categories_of_law.rb
@@ -1,6 +1,7 @@
 require_relative './categories_of_law/discrimination'
 require_relative './categories_of_law/education'
-require_relative './categories_of_law/immigration_and_asylum'
+require_relative './categories_of_law/asylum'
+require_relative './categories_of_law/immigration'
 
 module SubmissionConfig
   module CategoriesOfLaw
