--- lib/submission_config/categories_of_law/immigration_and_asylum.rb
@@ -1,9 +1,0 @@
-require_relative './immigration_and_asylum/immigration'
-require_relative './immigration_and_asylum/asylum'
-
-module SubmissionConfig
-  module CategoriesOfLaw
-    module ImmigrationAndAsylum
-    end
-  end
-end
